added Params to dict to libcli
