+++ src/haddock/libs/libcli.py
@@ -1,5 +1,6 @@
 """Add functionalities for CLIs."""
 from argparse import ArgumentTypeError
+from argparse import Action
 from functools import partial
 from pathlib import Path
 
@@ -68,3 +69,40 @@ def add_output_dir_arg(ap):
         type=Path,
         default=Path.cwd(),
         )
+
+
+class _ParamsToDict(Action):
+    """
+    Convert command-line parameters in an argument to a dictionary.
+
+    Example
+    -------
+
+    Where ``-x`` is an optional argument of the command-line client
+    interface.
+
+        >>> par1 1 par2 'my name' par3 [1,2,3] par4 True
+        >>> {'par1': 1, 'par2': 'my name', 'par3': [1, 2, 3]}
+
+    """
+
+    def __call__(self, parser, namespace, ivalues, option_string=None):
+        """Execute."""
+        params = ivalues[::2]
+        values = ivalues[1::2]
+
+        if len(params) != len(values):
+            raise parser.error(
+                "The parameters and value pairs "
+                "do not match for argument `-p`"
+                )
+
+        param_dict = {}
+        for k, v in zip(params, values):
+            print(f"k {k} v {v}")
+            try:
+                param_dict[k] = v
+            except (ValueError, TypeError, SyntaxError):
+                raise parser.error(f"Parameter {k} with invalid value {v}")
+
+        setattr(namespace, self.dest, param_dict)
